Texture specs can now have a NULL char set spec.

+++ source/graphics/2d/Textures/Char/CharSet.c
@@ -99,7 +99,7 @@ uint8 CharSet::getUsageCount()
  *
  * @return				Allocation type
  */
-uint32 CharSet::getAllocationType()
+uint8 CharSet::getAllocationType()
 {
 	return this->charSetSpec->allocationType;
 }
